+++ NEWS.md
@@ -2,6 +2,7 @@
 * Updated plotting modes to be one of : "re", "im", "mod" or "arg".
 * Updated int_spec function to use "re", "im", or "mod".
 * Added a function to replicate data across a particular dimension.
+* Added a convenience function to simulate normal looking 1H brain MRS data.
 
 # spant 0.7.0
 * Added new function for frequency drift correction.